Republish Edition after rendering PDF
We currently render PDFs after publishing the HTML attachment, because
we need to fetch the fully rendered HTML attachment from government-
frontend after publishing. We use this to create the PDF.

Since the PDF is generated after publication, it’s not listed on the
government-frontend page until the Edition is republished.