Add LocalisationMiddleware
Whitehall currently patches the Rails router to support localised
requests with the locale appended to the url e.g.
`gov.uk/government/announcements.fr`.

This is complex and is preventing us from upgrading to Rails 5.

This commit adds a Rack middleware that provides this functionality
which will allow us to retire the patch.