Fix content_type
The previous version of Carrierwave seemed to be setting the content
type of uploaded files when the browswer didn't specifically set it. The
upgraded version doesn't seem to do this although it's not clear whether
this is the way the tests are presenting files with fixture_file_upload
or Carrierwave itself. This commit manually sets the content_type using
the mime-types gem which preserves the old version behaviour. Once the
rails 5 upgrade is complete we should revisit this.