Fix attachable test
Test was failing since carrierwave has been updated as
`CarrierWave::SanitizedFile#blank?` now seems to check for the presence
of the file on the disk and the simulated virus scan in the factory
moves the file without updating the uploader path.

This commit also fixes some deprecation warnings about reloading
association collections.