Fix Classification test
Classification is labelled as an abstract class and doesn't get sent to
publishing API. Rails 5 changes that now fire `after_commit` callbacks
have highlighted that we were previously testing this behaviour. The
test began to fail when Rails 5 happened.

This commit removes the test and adds an equivalent unit test to the
sub-classes (`Topic` and `TopicalEvent`).

Also renames the test file to follow Rails conventions.