Add disable_publishes_to_publishing_api helper
We implemented the `PublishesToPublishingApi` module to queue up its
jobs in an `after_commit` callback. Up until Rails 5 these weren't
'fired' within tests as the tests used transactional fixtures.

As of Rails 5 transactional tests now fire these callbacks so many of
the tests that assert requests to publishing api are now failing as
there are extra requests being made during test set up, often from
within seemingly unrelated factories.

To compound matters a lot of the expectations for these requests are
made using helpers in gds-api-adapters which currently don't accept any
'times' arguments so updating the expectations to include the extra
requests is slightly more complicated (and a bad solution in any case).

Changing `PublishesToPublishingApi` to be a discreet object that saves
a supplied object and then queues up a job would be the best solution
but would involve quite a lot of changes.

As part of the work to upgrade Whitehall to Rails 5 (since 4 is no
longer supported), we need to get these tests to pass until the better
solution can be implemented.

This commit adds a helper that can be used to suppress the
`PublishesToPublishingApi` calls within a block.