Remove empty directory when uploading assets to Asset Manager
We noticed that the worker was only deleting the file even though the
storage engine was creating the file in its own unique directory. This
change means that the directory is also deleted which avoids us ending
up with lots of empty directories.

We're using `Dir.mktmpdir` in our test set-up so that we have a
directory that we have permission to remove. The `Tempfile` is created
in `Dir.tmpdir` by default but we don't have permission to delete that
and so the test we added would fail.